move: components > common> Auth 폴더 위치 components > Auth로 변경
